fix(evals): address Dileep's second review — config on add, version-switch dirty, model priority, composite race, shared buildRunConfig
