+++ frontend/src/sections/common/EvalPicker/serializeEvalConfig.js
@@ -31,6 +31,41 @@ export function sanitizeEvalMapping(mapping) {
   );
 }
 
+export const isCompositeEval = (evalConfig) =>
+  evalConfig.templateType === "composite" ||
+  evalConfig.evalTemplate?.template_type === "composite" ||
+  evalConfig.evalTemplate?.templateType === "composite";
+
+export const buildRunConfig = (evalConfig) => {
+  const runConfig = {};
+  if (!isCompositeEval(evalConfig)) {
+    if (evalConfig.model) runConfig.model = evalConfig.model;
+    if (evalConfig.agent_mode) runConfig.agent_mode = evalConfig.agent_mode;
+    if (evalConfig.check_internet !== undefined)
+      runConfig.check_internet = !!evalConfig.check_internet;
+    if (evalConfig.summary) runConfig.summary = evalConfig.summary;
+    if (evalConfig.knowledge_base_id)
+      runConfig.knowledge_base_id = evalConfig.knowledge_base_id;
+    if (evalConfig.knowledge_bases)
+      runConfig.knowledge_bases = evalConfig.knowledge_bases;
+    if (evalConfig.tools) runConfig.tools = evalConfig.tools;
+    if (evalConfig.pass_threshold !== undefined)
+      runConfig.pass_threshold = evalConfig.pass_threshold;
+    if (
+      evalConfig.choice_scores &&
+      Object.keys(evalConfig.choice_scores).length
+    )
+      runConfig.choice_scores = evalConfig.choice_scores;
+    if (evalConfig.multi_choice !== undefined)
+      runConfig.multi_choice = !!evalConfig.multi_choice;
+  }
+  if (evalConfig.data_injection)
+    runConfig.data_injection = evalConfig.data_injection;
+  if (evalConfig.error_localizer_enabled !== undefined)
+    runConfig.error_localizer_enabled = !!evalConfig.error_localizer_enabled;
+  return runConfig;
+};
+
 export function serializeEvalConfig(evalConfig) {
   const runConfig = {};
   for (const k of RUN_CONFIG_KEYS) {
