+++ frontend/src/sections/common/EvaluationDrawer/EvaluationDrawer.jsx
@@ -40,6 +40,7 @@ import CreateEvaluationGroupDrawer from "./CreateEvaluationGroupDrawer";
 import { useSearchParams } from "react-router-dom";
 import { resetEvalStore } from "src/sections/evals/store/useEvalStore";
 import { EvalPickerDrawer } from "src/sections/common/EvalPicker";
+import { buildRunConfig } from "src/sections/common/EvalPicker/serializeEvalConfig";
 import { useRunEvaluationStore } from "src/sections/develop-detail/states";
 import { useWorkbenchEvaluationContext } from "src/sections/workbench/createPrompt/Evaluation/context/WorkbenchEvaluationContext";
 
@@ -594,42 +595,7 @@ const EvaluationDrawerChild = ({
           // should be forwarded so eval_runner can apply it at run time.
           const isComposite = evalConfig.templateType === "composite";
 
-          const runConfig = {};
-          if (!isComposite) {
-            // Single-eval runtime overrides. Composite children each
-            // carry their own model/mode/tools — none of this applies
-            // at the composite binding level.
-            if (evalConfig.model) runConfig.model = evalConfig.model;
-            if (evalConfig.agent_mode)
-              runConfig.agent_mode = evalConfig.agent_mode;
-            if (evalConfig.check_internet !== undefined)
-              runConfig.check_internet = !!evalConfig.check_internet;
-            if (evalConfig.summary) runConfig.summary = evalConfig.summary;
-            if (evalConfig.knowledge_base_id)
-              runConfig.knowledge_base_id = evalConfig.knowledge_base_id;
-            if (evalConfig.knowledge_bases)
-              runConfig.knowledge_bases = evalConfig.knowledge_bases;
-            if (evalConfig.tools) runConfig.tools = evalConfig.tools;
-            if (evalConfig.pass_threshold !== undefined)
-              runConfig.pass_threshold = evalConfig.pass_threshold;
-            if (
-              evalConfig.choice_scores &&
-              Object.keys(evalConfig.choice_scores).length
-            )
-              runConfig.choice_scores = evalConfig.choice_scores;
-            if (evalConfig.multi_choice !== undefined)
-              runConfig.multi_choice = !!evalConfig.multi_choice;
-          }
-          // Data injection applies to both single and composite — the
-          // backend resolves it at row-evaluation time.
-          if (evalConfig.data_injection)
-            runConfig.data_injection = evalConfig.data_injection;
-          // Error localizer toggle was previously dropped between
-          // EvalPickerConfigFull and the backend. It now flows through
-          // for both single and composite bindings.
-          if (evalConfig.error_localizer_enabled !== undefined)
-            runConfig.error_localizer_enabled =
-              !!evalConfig.error_localizer_enabled;
+          const runConfig = buildRunConfig(evalConfig);
 
           // Code-eval static params (function_params_schema values).
           // `EvalPickerConfigFull.handleSave` hands them back on
@@ -665,9 +631,13 @@ const EvaluationDrawerChild = ({
               name: evalConfig.name,
               template_id: evalConfig.templateId,
               model: isComposite ? undefined : evalConfig.model,
+              // Optimizer runs evals itself; dataset adds are save-only so
+              // an add doesn't queue work the user didn't ask for.
               run: module !== "run-optimization" && module !== "dataset",
+              // Surface error_localizer at the top level so
+              // EditAndRunUserEvalView can update eval_metric.error_localizer.
               error_localizer: runConfig.error_localizer_enabled ?? false,
-              ...(evalConfig.isDirty
+              ...(!evalConfig.userEvalId || evalConfig.isDirty
                 ? {
                     config: {
                       mapping: evalConfig.mapping || {},
@@ -681,15 +651,12 @@ const EvaluationDrawerChild = ({
                     },
                   }
                 : {}),
-              ...(isComposite && evalConfig.compositeWeightOverrides
+              ...(isComposite &&
+              (evalConfig.compositeWeightOverrides ??
+                evalConfig.composite_weight_overrides)
                 ? {
                     composite_weight_overrides:
-                      evalConfig.compositeWeightOverrides,
-                  }
-                : {}),
-              ...(isComposite && evalConfig.composite_weight_overrides
-                ? {
-                    composite_weight_overrides:
+                      evalConfig.compositeWeightOverrides ??
                       evalConfig.composite_weight_overrides,
                   }
                 : {}),
@@ -713,6 +680,8 @@ const EvaluationDrawerChild = ({
               effectiveModule === "experiment")
           ) {
             try {
+              // `id` is always the datasetId for dataset/experiment flows
+              // (experiment evals still live under a dataset).
               await axios.post(
                 endpoints.develop.eval.editEval(id, evalConfig.userEvalId),
                 {
